+++ releases.md
@@ -1,5 +1,9 @@
 # Releases
 
+## Unreleased
+
+  - Add `Memory::Sampler#as_json` and `#to_json`.
+
 ## v0.6.0
 
   - Add agent context.